move new logging files to logging2 package
