+++ src/test/java/org/jitsi/utils/logging2/BaseLoggerTest.java
@@ -1,4 +1,4 @@
-package org.jitsi.utils.logging;
+package org.jitsi.utils.logging2;
 
 import org.junit.*;
