+++ src/test/java/org/jitsi/utils/logging2/LogContextTest.java
@@ -1,5 +1,6 @@
-package org.jitsi.utils.logging;
+package org.jitsi.utils.logging2;
 
+import org.jitsi.utils.logging2.*;
 import org.junit.*;
 
 import java.util.*;
